schemex v0.1.0 Schemex

Small Elixir wrapper for Confluent Schema Registry.

Available methods:

config/1

subjects/1

versions/2

version/3

latest/2

schema/2

delete/2

delete/3

register/3

check/3

test/4

update_compatibility/2

update_compatibility/3

Functions

check(host, subject, schema)

Check whether a schema has been registered under subject “value”

  Schemex.check("http://localhost:8081", "value", %{"type" => "string"})

config(host)

Get top level config.

  Schemex.config("http://localhost:8081")

delete(host, subject)

Delete all versions of the schema registered under subject “value”

  Schemex.delete("http://localhost:8081", "value")

delete(host, subject, version)

Delete version 1 of the schema registered under subject “value”

  Schemex.delete("http://localhost:8081", "value", 1)

latest(host, subject)

Fetch latest version of the schema registered under subject “value”

  Schemex.latest("http://localhost:8081", "value")

register(host, subject, schema)

Register a new version of a schema under the subject “Kafka-key”

  Schemex.register("http://localhost:8081", "value", %{"type" => "string"})

schema(host, id)

Fetch a schema by globally unique id 1

  Schemex.schema("http://localhost:8081", 1)

subjects(host)

List all subjects.

  Schemex.subjects("http://localhost:8081")

test(host, subject, schema, version \\ "latest")

Test compatibility of a schema with specific version of schema under subject “value”

  Schemex.test("http://localhost:8081", "value", %{"type" => "string"})

update_compatibility(host, comp)

Update compatibility requirements globally

  Schemex.update_compatibility("http://localhost:8081", "NONE")

update_compatibility(host, subject, comp)

Update compatibility requirements under the subject “value”

  Schemex.update_compatibility("http://localhost:8081", "value", "NONE")

version(host, subject, version)

Fetch version 1 of the schema registered under subject “value”

  Schemex.version("http://localhost:8081", "value", 1)

versions(host, subject)

List all schema versions registered under the subject “value”

  Schemex.versions("http://localhost:8081", "value")